refactor: UI 구조 재구성
mapping UI 개발 시 길 안내 화면과 동일하게 카메라 화면을 보면서 제작하도록 만들기 위해 GLSurfaceView 를 main activity로 옮김.
